recombined saveh5 and loadh5 to just h5
