+++ dnplab/dnpIO/__init__.py
@@ -5,8 +5,7 @@ from . import vnmrj
 from . import specman
 from . import bes3t
 from . import parspc
-from . import saveh5
-from . import loadh5
+from . import h5
 from . import power
 from . import cnsi
 from . import vna
